==> scripts_dia/scripts/01-Tratamento_Dados_v1.0.ipynb <==
## Descrição do Notebook

> Este é o notebook ao qual o **conjunto de dados foi utilizado** para a expansão dos dados para os próximos anos

import pandas as pd
import xlrd
import matplotlib.pyplot as plt
from datetime import datetime
import numpy as np
import seaborn as sns

# Lendo o arquivo excel com a sheet name 2158151
df = pd.read_excel("../datasets/base_dados.xlsx", sheet_name = "2158151")
# Lendo o arquivo excel com todas as sheet names
df_cheio =  pd.read_excel("../datasets/base_dados.xlsx", sheet_name = None)

# Função para mapear as colunas do dataframe pelo separador ';'
def trata_df(dataframe):
    df = dataframe.iloc[:, 0].str.split(';', expand=True)
    df.columns = [n.replace('"', '') for n in  dataframe.columns.str.split(';')[0]]
    return df

for i in range(len(list(df_cheio.keys()))-1):
    chaves = list(df_cheio.keys())
    df = pd.concat([df, df_cheio[chaves[i+1]]])

# Fazendo o tratamento nas datas do dataframe    
df_tratado = trata_df(df)
data = [xlrd.xldate_as_datetime(int(nume), 0).strftime("%d-%m-%Y") for nume in df_tratado["DATA_TRAT."]]
df_tratado['DATA_TRAT.'] = data

df_tratado.head()

Transformando as colunas para tipos numéricos e formatação numérica padrão estadunidense.

def isfloat(str):
    try: 
        float(str)
    except ValueError: 
        return False
    return True

df_tratado = df_tratado.loc[df_tratado["Preço"].str.replace(",", ".").apply(isfloat)]
df_tratado["Preço"] = df_tratado["Preço"].str.replace(",", ".").astype("float")

df_tratado["Vendas Reais Qtde"] = df_tratado["Vendas Reais Qtde"].astype("float")
df_tratado["Vendas Reais $"] = df_tratado["Vendas Reais $"].str.replace(",", ".").astype("float")
df_tratado["CMV"] = df_tratado["CMV"].str.replace(",", ".").astype("float")
df_tratado["preço2"] = df_tratado["preço2"].str.replace(",", ".").astype("float")
df_tratado["CMV/QTD"] = df_tratado["CMV/QTD"].str.replace(",", ".").astype("float")
df_tratado["Mg %"] = df_tratado["Mg %"].str.replace(",", ".").astype("float")
df_tratado["Margem Líquida Real %"] = df_tratado["Margem Líquida Real %"].str.replace(",", ".").astype("float")
df_tratado["Margem Líquida Real $"] = df_tratado["Margem Líquida Real $"].str.replace(",", ".").astype("float")
df_tratado["Venda Líquida Real $"] = df_tratado["Venda Líquida Real $"].str.replace(",", ".").astype("float")





Filtrando apenas as colunas que possuem valor maior que 0. Dessa forma, não teremos itens que não foram vendidos e nem itens com preço igual a 0.

num_cols = ["Vendas Reais Qtde", "Vendas Reais $", "CMV", "Margem Líquida Real %", "Margem Líquida Real $", 
           "Venda Líquida Real $", "Preço", "preço2", "CMV/QTD", "Mg %"]
for col in num_cols:
    df_tratado = df_tratado.loc[df_tratado[col] > 0]

df_tratado = df_tratado.reset_index(drop = True)
df_tratado.head()

Gerando o **dataset** final.

- Removendo preços que não variaram ao longo de 2017

lst_uf = df_tratado["UF"].unique()
lst_sap = df_tratado["SAP+NOME"].unique()

for uf in lst_uf:
    for sap in lst_sap:
        df_seg = df_tratado.loc[(df_tratado["UF"] == uf) & (df_tratado["SAP+NOME"] == sap) & 
                                (df_tratado["DATA_TRAT."] <= '2017-12-31')].copy()
        if df_seg.empty == True:
               continue
        else:
            if len(df_seg["Preço"].unique()) == 1:
                #print(f'{uf} - {sap}')
                ind = df_seg['Preço']
                df_tratado = df_tratado.drop(ind.index)

df_tratado.to_csv("../datasets/df_rpo_v2.csv", index = False)
