==> scripts_dia/scripts/02-Expansão_de_Dados_v2.ipynb <==
### Descrição do Notebook

Notebook de expansão de dados. Neste notebook foi aplicada a seguinte lógica:
1. Em **2018**, cada item e UF foi aplicada uma varição de preços atráves de um vetor aleatório ponderado onde o preço poderia se manter igual ou aumentar 1% ou 2%. A probabilidade ponderada para cada um dos preços é 60%, 20%, 20%, respectivamente.

2. Em **2019**, cada item e UF foi aplicada uma varição de 5% e, de maneira similar, e, em média, 50% dos produtos se mantiveram com o mesmo preço e 50% dos outros produtos sofreram o aumento.

import pandas as pd
import random
from itertools import product

# Gerando dados de 2018

df2017 = pd.read_csv('../datasets/df_rpo_v2.csv')
df2017["DATA_TRAT."] = pd.to_datetime(df2017["DATA_TRAT."], dayfirst = True)
df2017 = df2017.loc[df2017["DATA_TRAT."] <= "2017-12-31"].copy()
df2017 = df2017.reset_index(drop = True)

df2018 = df2017.copy()

df2018['DATA_TRAT.'] = pd.to_datetime(df2018['DATA_TRAT.'])

Código que percorre produto e UF e aplica a regra de aumento de preços:

lista_produtos  = df2018["SAP+NOME"].unique()
lista_uf = df2018["UF"].unique()

for produto in lista_produtos:
    for uf in lista_uf:
        vetor = [1, 1.01, 1.02]
        colunas = ["Preço", "Vendas Reais $", "CMV", "Margem Líquida Real %", "Margem Líquida Real $", "Mg %"]
        df_aux = df2018.loc[(df2018["SAP+NOME"] == produto) & (df2018["UF"] == uf)].copy()
               
        for index, row in df_aux[colunas].iterrows():
            random.seed(50)
            n_ = random.choices(vetor, weights = [60, 20, 20])
            df2018.iloc[index, [9, 4, 5, 6, 7, 12]] = row[colunas] * n_

Substituindo as datas para o ano seguinte.

data = df2018['DATA_TRAT.']
data = data.apply(lambda x: x.replace(year=2018))
df2018['DATA_TRAT.'] = data

df2018 = df2018.sort_values(by=['SAP+NOME', 'UF', 'DATA_TRAT.'], ascending=True)

df2018.to_csv('../datasets/df2018.csv', index=False)

# Gerando dados de 2019

df2019 = df2018.copy()

df2019['DATA_TRAT.'] = pd.to_datetime(df2019['DATA_TRAT.'])

data = df2019['DATA_TRAT.']
data = data.apply(lambda x: x.replace(year=2019))
df2019['DATA_TRAT.'] = data

Código que percorre produto e UF e aplica a regra de aumento de preços:

lista_produtos  = df2019["SAP+NOME"].unique()
lista_uf = df2019["UF"].unique()

for produto in lista_produtos:
    for uf in lista_uf:
        vetor = [1, 1.05]
        colunas = ["Preço", "Vendas Reais $", "CMV", "Margem Líquida Real %", "Margem Líquida Real $", "Mg %"]
        df_aux = df2018.loc[(df2019["SAP+NOME"] == produto) & (df2019["UF"] == uf)].copy()
        
        
        for index, row in df_aux[colunas].iterrows():
            random.seed(50)
            n_ = random.choices(vetor, weights = [50, 50])
            df2019.iloc[index, [9, 4, 5, 6, 7, 12]] = row[colunas] * n_

df2019 = df2019.sort_values(by=['SAP+NOME', 'UF', 'DATA_TRAT.'], ascending=True)
df2019.head(30)

df2019.to_csv('../datasets/df2019.csv', index=False)

# Gerando DF final com os 3 anos

Concatenando os dataframes gerados anteriormente.

dfFinal = pd.concat([df2017, df2018, df2019]).reset_index(drop=True)
dfFinal['DATA_TRAT.'] = pd.to_datetime(dfFinal['DATA_TRAT.'], format='%Y-%m-%d')
dfFinal = dfFinal.sort_values(by=['SAP+NOME', 'UF', 'DATA_TRAT.'], ascending=True)

dfFinal.to_csv('../datasets/df_rpo_v3.csv', index=False)
