==> scripts_dia/scripts/05-Tratamento_de _dados_para_modelagem.ipynb <==
## Descrição do Notebook

Este é um notebook somente para **EDA**. Os tratamentos feitos aqui foram feitos sobre os dados brutos iniciais e o *dataset* final gerado não foi utilizado como base na expansão dos dados.

import pandas as pd
import xlrd

from fuzzywuzzy import process
import numpy as np
import datetime

### Leitura  e observação do df principal

df_tratado = pd.read_csv('../datasets/df_rpo_v4_dia.csv')
df_tratado.shape

## Filtrar ano de 2017

df_tratado = df_tratado.loc[df_tratado["Ano"]==2017]

df_tratado

### Checagem de duplicidade no df principal

df_tratado

df_tratado.duplicated(subset=['SAP+NOME', 'UF', 'Ano', 'Mes', "Dia"]).value_counts()

### Criação da coluna da variável dummy de preço psicologico

df_tratado['DPPsic'] = ''
df_tratado.loc[df_tratado['Preço'] != df_tratado['preço2'], 'DPPsic'] = 0
df_tratado.loc[df_tratado['Preço'] == df_tratado['preço2'], 'DPPsic'] = 1

### Calculo de correlação entre quantidade de vendas e preço

df_corr = df_tratado.groupby(['UF', 'SAP+NOME'])[['Preço', 'Vendas Reais Qtde']].corr().reset_index()

df_corr

df_corr.isnull().sum()

df_corr = df_corr.loc[(df_corr['level_2'] == 'Vendas Reais Qtde') & (df_corr['Preço'].notnull()),\
                      ['UF', 'SAP+NOME','level_2','Preço']].sort_values(by='Preço', ascending=True).reset_index()

df_corr = df_corr[['UF', 'SAP+NOME', 'Preço']]

df_corr

### Cálculo do numero de observações por série

df_corr_obs = df_tratado.groupby(['UF', 'SAP+NOME'])['Dia'].count().reset_index()

df_corr_obs

### DF com correlação entre preço e quantidade e número de dias observadas por série

df_corr_final = pd.merge(df_corr_obs, df_corr, how='left', on=['UF', 'SAP+NOME'])
df_corr_final.rename(columns={"Preço":"Correlacao", "Dia":'NObs'}, inplace = True)
df_corr_final = df_corr_final.sort_values('NObs', ascending=False)
df_corr_final.reset_index(drop=True)

### CSV para dashboard
#df_corr_final.to_csv('../datasets/db_corr_nobs.csv', decimal = ',', sep=';', index=False)

### Acrescenta o número de semanas observadas ao df principal, renomeia SAP+NOME

df_tratado_final = pd.merge(df_tratado, df_corr_final[['UF', 'SAP+NOME', 'NObs']], 
                            how='left', left_on=['UF', 'SAP+NOME'], right_on=['UF', 'SAP+NOME'])

df_tratado_final.rename(columns={'SAP+NOME':'Produto', 'Vendas Reais Qtde':'QTD'}, inplace = True)

sap_nome = df_tratado_final['Produto'].str.split(" - ", n=1, expand=True)

df_tratado_final['SAP'] = sap_nome[0].apply(lambda x: x.lstrip('0'))

df_tratado_final['Produto'] = sap_nome[1]

df_tratado_final

# df_tratado_final.to_csv('../datasets/df_tratado_final.csv', decimal = '.', sep=';', index=False)

### Criando df que será utilizado para ajustar o modelo - Produto individualizado

df_modelo_pi = df_tratado_final.loc[df_tratado_final['NObs'] >= 16, ['SAP', 'Produto', 'UF', 'Ano', 'Mes', 'Dia',
                                                                   'Preço', 'QTD', 'CMV/QTD',
                                                                   'DPPsic']]

df_modelo_pi

# SORTED()

aux = df_modelo_pi.copy()

lista_unicos = sorted(list(aux["Produto"].unique()))
aux["GrpProd"] = np.nan
grupo = 0
thresh = 52

# While

while len(lista_unicos) > 0:
    item = lista_unicos[0] 
    similaridades = process.extract(item, lista_unicos)
    for i in range(len(similaridades)):
        if (similaridades[i][1] >= thresh):
            aux.loc[aux["Produto"] == similaridades[i][0], "GrpProd"] = grupo
            lista_unicos.remove(similaridades[i][0])
    grupo += 1

## Validação

for i in range(len(aux["GrpProd"].unique())):
    print(f'grupo => {i}')
    print(aux.loc[aux["GrpProd"] == i]["Produto"].unique())

sorted(aux["GrpProd"].unique())[-1]

## Realizando os agrupamentos

def agrupamentos(dataframe, threshold:float):
    """
    Descreve uma função que classifica os itens em grupos de similaridade utilizando fuzzy matching.
    
    dataframe -> dataframe que contém os itens a serem classificados]
    threshold -> grau de similaridade entre os nomes dos itens (0 a  100)
    """
    
    # Cria um vetor de itens únicos ordenados alfabeticamente
    lista_unicos = sorted(list(dataframe["Produto"].unique()))
    
    # Cria uma feature de NaN's chamada Grupo
    dataframe["GrpProd"] = np.nan
    grupo = 0
    
    # Percorre o nosso dataframe e faz o agrupamento entre os itens
    while len(lista_unicos) > 0:
        item = lista_unicos[0] 
        similaridades = process.extract(item, lista_unicos)
        for i in range(len(similaridades)):
            if (similaridades[i][1] >= threshold):
                dataframe.loc[dataframe["Produto"] == similaridades[i][0], "GrpProd"] = grupo
                lista_unicos.remove(similaridades[i][0])
        grupo += 1
    
    return dataframe

### Criando df que será utilizado para ajustar o modelo - Produto agrupado

df_modelo_pa = agrupamentos(aux, 55)

aux['PrXQtd'] = aux['Preço'] * aux['QTD'] 

aux['CMVXQtd'] = aux['CMV/QTD'] * aux['QTD']

aux['PPsicXQtd'] = aux['DPPsic'] * aux['QTD']

aux['PPsicXQtd'] = aux['PPsicXQtd'].astype(int)

df_modelo_pa = aux[['GrpProd', 'UF', 'Ano', 'Mes', 'Dia', 'QTD', 
                   'PrXQtd', 'CMVXQtd', 'PPsicXQtd']].groupby(['GrpProd', 'UF', 'Ano','Mes','Dia']).sum().reset_index()
          
df_modelo_pa['Preço'] = df_modelo_pa['PrXQtd'] / df_modelo_pa['QTD']

df_modelo_pa['CMV/QTD'] = df_modelo_pa['CMVXQtd'] / df_modelo_pa['QTD']

df_modelo_pa['DPPsic'] = df_modelo_pa['PPsicXQtd'] / df_modelo_pa['QTD']

df_modelo_pa.drop(columns=['CMVXQtd', 'PrXQtd', 'PPsicXQtd'], inplace = True)

df_modelo_pa.head(60)

df_modelo_pa.to_csv('../datasets/df_modelo_pa_diario.csv', decimal = '.', sep=';', index=False)

df_modelo_pi.to_csv('../datasets/df_modelo_pi_diario.csv', decimal = '.', sep=';', index=False)
